Document entry as <TermId, TF>

+++ include/utils.hpp
@@ -7,4 +7,6 @@ using TermId_t = uint32_t;
 
 // Store documents as vector of (TermId, TF) pairs
 using DocId_t = uint32_t;
-using Document_t = std::vector<std::pair<TermId_t, int>>;
+using Document_t = std::vector<std::pair<
+
+    TermId_t, float>>;
